fix: extract SearchResult into vector_types.py to break qdrant-client import chain (#447)
Creates zero-dependency vector_types.py with SearchResult dataclass so that
rag_service, forms_query_service, and vector_chroma can import it without
pulling qdrant-client at startup. Removes the duplicate definition from
pgvector_service.py and updates __init__.py re-exports.

Unblocks #449 (qdrant-client removal from requirements).
